requirements.txt
numpy==2.1.1
onnxruntime==1.19.2
coloredlogs==15.0.1
flatbuffers==24.3.25
humanfriendly==10.0
mpmath==1.3.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sympy==1.13.3
tzdata==2024.2
